+++ package.json
@@ -11,12 +11,12 @@
 	},
 	"workspaces": ["apps/*", "packages/*", "."],
 	"scripts": {
-		"build": "npm run build:build-config && tsup",
-		"dev": "tsup --watch='packages/*/src/**/*' --watch='apps/cli/src/**/*' --watch='bin/**/*' --watch='mcp-server/**/*'",
+		"build": "npm run build:build-config && tsdown",
+		"dev": "tsdown --watch='packages/*/src/**/*' --watch='apps/cli/src/**/*' --watch='bin/**/*' --watch='mcp-server/**/*'",
 		"turbo:dev": "turbo dev",
 		"turbo:build": "turbo build",
 		"turbo:typecheck": "turbo typecheck",
-		"dev:main": "tsup --watch --onSuccess 'echo \"📦 Main package built\" && npm link'",
+		"dev:main": "tsdown --watch --onSuccess 'echo \"📦 Main package built\" && npm link'",
 		"dev:legacy": "npm run build:build-config && concurrently -n \"core,cli,main\" -c \"blue,green,yellow\" \"npm run dev:core\" \"npm run dev:cli\" \"npm run dev:main\"",
 		"dev:core": "npm run dev -w @tm/core",
 		"dev:cli": "npm run dev -w @tm/cli",
@@ -143,7 +143,7 @@
 		"mock-fs": "^5.5.0",
 		"prettier": "^3.5.3",
 		"supertest": "^7.1.0",
-		"tsup": "^8.5.0",
+		"tsdown": "^0.15.2",
 		"tsx": "^4.16.2",
 		"turbo": "^2.5.6",
 		"typescript": "^5.9.2"
