+++ packages/build-config/src/tsup.base.ts
@@ -1,8 +1,8 @@
 /**
- * Base tsup configuration for Task Master monorepo
+ * Base tsdown configuration for Task Master monorepo
  * Provides shared configuration that can be extended by individual packages
  */
-import type { Options } from 'tsup';
+import type { UserConfig } from 'tsdown';
 
 const isProduction = process.env.NODE_ENV === 'production';
 const isDevelopment = !isProduction;
@@ -17,85 +17,28 @@ export const env = {
 };
 
 /**
- * Base tsup configuration for all packages
+ * Base tsdown configuration for all packages
  * Since everything gets bundled into root dist/ anyway, use consistent settings
  */
-export const baseConfig: Partial<Options> = {
-	format: ['esm'],
-	target: 'node18',
+export const baseConfig: Partial<UserConfig> = {
 	sourcemap: isDevelopment,
-	clean: true,
-	dts: false,
+	dts: isDevelopment,
 	minify: isProduction,
 	treeshake: isProduction,
-	splitting: false,
 	// Don't bundle any other dependencies (auto-external all node_modules)
-	external: [/^[^./]/],
-	esbuildOptions(options) {
-		options.platform = 'node';
-		// Allow importing TypeScript from JavaScript
-		options.resolveExtensions = ['.ts', '.js', '.mjs', '.json'];
-		// Better source mapping in development only
-		options.sourcesContent = isDevelopment;
-		// Keep original names for better debugging in development
-		options.keepNames = isDevelopment;
-	},
-	// Watch mode configuration for development
-	watch: false
+	external: [/^[^./]/]
 };
 
 /**
- * Legacy external modules list - kept for backwards compatibility
- * Note: When using tsup-node, this is not needed as it automatically
- * excludes dependencies and peerDependencies from package.json
- */
-export const commonExternals = [
-	// Native Node.js modules (for cases where tsup is used instead of tsup-node)
-	'fs',
-	'path',
-	'child_process',
-	'crypto',
-	'os',
-	'url',
-	'util',
-	'stream',
-	'http',
-	'https',
-	'events',
-	'assert',
-	'buffer',
-	'querystring',
-	'readline',
-	'zlib',
-	'tty',
-	'net',
-	'dgram',
-	'dns',
-	'tls',
-	'cluster',
-	'process',
-	'module'
-];
-
-/**
  * Utility function to merge configurations
- * Simplified for tsup-node usage
+ * Simplified for tsdown usage
  */
 export function mergeConfig(
-	baseConfig: Partial<Options>,
-	overrides: Partial<Options>
-): Options {
+	baseConfig: Partial<UserConfig>,
+	overrides: Partial<UserConfig>
+): UserConfig {
 	return {
 		...baseConfig,
-		...overrides,
-		// Merge esbuildOptions
-		esbuildOptions(options, context) {
-			if (baseConfig.esbuildOptions) {
-				baseConfig.esbuildOptions(options, context);
-			}
-			if (overrides.esbuildOptions) {
-				overrides.esbuildOptions(options, context);
-			}
-		}
-	} as Options;
+		...overrides
+	} as UserConfig;
 }
